Product Service With Exception Handling

+++ product-service/src/main/java/com/zensar/demo/service/ProductServiceImpl.java
@@ -11,6 +11,8 @@ import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import com.zensar.demo.dto.ProductDto;
 import com.zensar.demo.entity.Product;
+import com.zensar.demo.exception.EmptyProductList;
+import com.zensar.demo.exception.ProductNotFound;
 import com.zensar.demo.repository.ProductRepository;
 
 @Service
@@ -21,12 +23,13 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Autowired
 	private ModelMapper modelMapper;
-	
+
 	Pageable p;
 
 	@Override
 	public ProductDto getProduct(int productId) {
-
+		if (!productRepository.findById(productId).isPresent())
+			throw new ProductNotFound();
 		Product product = productRepository.findById(productId).get();
 		return modelMapper.map(product, ProductDto.class);
 
@@ -35,6 +38,8 @@ public class ProductServiceImpl implements ProductServices {
 	@Override
 	public List<ProductDto> getAllProduct(int pageNumber, int pageSize, String sortBy, Direction direction) {
 
+		if (productRepository.findAll().isEmpty())
+			throw new EmptyProductList();
 		List<ProductDto> listOfDto = new ArrayList<ProductDto>();
 		Page<Product> page = productRepository.findAll(PageRequest.of(pageNumber, pageSize, direction, sortBy));
 		for (Product product : page) {
@@ -46,6 +51,7 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public ProductDto insertProduct(ProductDto productDto) {
+		
 		Product product = modelMapper.map(productDto, Product.class);
 		Product product2 = productRepository.save(product);
 		return modelMapper.map(product2, ProductDto.class);
@@ -53,6 +59,9 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public void updateProduct(int productId, ProductDto productDto) {
+
+		if (!productRepository.findById(productId).isPresent())
+			throw new ProductNotFound();
 		Product product = modelMapper.map(productDto, Product.class);
 		Product product2 = productRepository.save(product);
 		modelMapper.map(product2, ProductDto.class);
@@ -61,12 +70,17 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public void deleteProduct(int productId) {
+		if (!productRepository.findById(productId).isPresent())
+			throw new ProductNotFound();
 		productRepository.deleteById(productId);
 
 	}
 
 	@Override
 	public List<ProductDto> findByProductName(String productName) {
+
+		if (productRepository.findByProductName(productName).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.findByProductName(productName);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
@@ -77,6 +91,8 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public List<ProductDto> findByProductCost(int productCost) {
+		if (productRepository.findByProductCost(productCost).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.findByProductCost(productCost);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
@@ -87,6 +103,9 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public List<ProductDto> findByProductCostBetween(int productCost1, int productCost2) {
+
+		if (productRepository.findByProductCostBetween(productCost1, productCost2).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.findByProductCostBetween(productCost1, productCost2);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
@@ -97,6 +116,8 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public List<ProductDto> test(String productName) {
+		if (productRepository.test(productName).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.test(productName);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
@@ -107,6 +128,8 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public List<ProductDto> test2(int productCost) {
+		if (productRepository.test2(productCost).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.test2(productCost);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
@@ -117,6 +140,8 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public List<ProductDto> test3(String productName, int productCost) {
+		if (productRepository.test3(productName, productCost).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.test3(productName, productCost);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
@@ -127,6 +152,9 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public List<ProductDto> test4(String productName, int productCost) {
+
+		if (productRepository.test4(productName, productCost).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.test4(productName, productCost);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
@@ -137,6 +165,8 @@ public class ProductServiceImpl implements ProductServices {
 
 	@Override
 	public List<ProductDto> findByProductCostGreaterThan(int productCost) {
+		if (productRepository.findByProductCostGreaterThan(productCost).isEmpty())
+			throw new EmptyProductList();
 		List<Product> listOfProduct = productRepository.findByProductCostGreaterThan(productCost);
 		List<ProductDto> lsitOfDto = new ArrayList<ProductDto>();
 		for (Product product : listOfProduct) {
